Initial Commit
This repo is created to separate the relatively large original documents that the static website of electronics museum is originated from so that cloning and working with its main repository becomes simpler and easier.

For preservation purposes, the QR Codes (and their corresponding txt files) still refer to the old website ( https://shamsipoor-museum.github.io/museum-website/ ) which is still maintained for the old printed QR Codes to continue to work, but will update in the follow up commit.